Initial Commit
All of the files regarding the project.

ROM contains the Chip-8 memory data to start games. 
CHIP contains the logic behind the CPUs innerworkings, including the registers, memory, program counter and the opcodes.
EMU contains the view of the Chip's results. To start, go to emu App, change the ROM to whichever you prefer (as long as it is inside the folder) and start.